# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kash-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/fx/rates": {
      "get": {
        "tags": [
          "fx"
        ],
        "operationId": "get_fx_rates",
        "parameters": [
          {
            "name": "from",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quotes",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "FX rates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFxRatesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Not logged in"
          },
          "500": {
            "description": "Internal server error"
          },
          "502": {
            "description": "Upstream FX provider error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExchangeRateRow": {
        "type": "object",
        "required": [
          "date",
          "currency",
          "rate"
        ],
        "properties": {
          "currency": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "rate": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "GetFxRatesResponse": {
        "type": "object",
        "required": [
          "rates"
        ],
        "properties": {
          "rates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExchangeRateRow"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "axum_session"
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ]
}